pyproject.toml
[build-system]
requires = ["maturin>=1.8"]
build-backend = "maturin"

[project]
name = "bookmarks"
version = "0.1.0"
description = "Agent-driven CLI for organizing bookmarks."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Adam Williams" }
]
dependencies = [
  "arize-phoenix>=12.25.1",
  "fastembed>=0.7.4",
  "huggingface-hub>=0.27.0",
  "langchain-community>=0.4",
  "langchain-core>=1.1.3",
  "langgraph>=1.0.4",
  "llama-cpp-python>=0.3.0",
  "nltk>=3.9",
  "opentelemetry-exporter-otlp>=1.39.1",
  "opentelemetry-sdk>=1.39.1",
  "smolagents>=1.22.0",
]

[project.scripts]
bookmarks = "bookmarks.cli:main"

[project.optional-dependencies]
dev = [
  "maturin>=1.8",
  "pytest>=9.0",
  "ruff>=0.14",
  "torch>=2.9"
]

[dependency-groups]
dev = [
  "maturin>=1.8",
  "pytest>=9.0",
  "ruff>=0.14",
  "torch>=2.9"
]

[tool.maturin]
module-name = "bookmarks._native"
bindings = "pyo3"
features = ["pyo3/extension-module", "pyo3/abi3-py311"]
python-source = "src"
